+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * **terra-dynamic-component-loader** is deprecated now. Use angular's [NgComponentOutlet](https://angular.io/api/common/NgComponentOutlet) instead.
 * **terra-toggle** is deprecated now. Use angular material's [slide-toggle](https://material.angular.io/components/slide-toggle) instead.
 * **terra-category-picker** removed provider for `TerraNodeTreeConfig` since it was not used anymore.
+* **terra-overlay** is deprecated now. Use angular material's [dialog](https://material.angular.io/components/dialog) instead.
+* **terra-input** is deprecated now. Use angular material's [input](https://material.angular.io/components/input) instead.
 
 # 5.3.0 (24.04.2020)